requirements.txt
redis==5.0.1
psycopg2-binary==2.9.9
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0